game starting tomorrow, chat done
