+++ app/models/user.py
@@ -36,6 +36,12 @@ class User(db.Model, UserMixin):
     def check_password(self, password):
         return check_password_hash(self.password, password)
 
+    def name(self):
+         return {
+            'id': self.id,
+            'username': self.username,
+         }
+
     def basic(self):
         return {
             'id': self.id,
